Moved PIN and size up tp /[pin]/+layout.server.js
