+++ src/routes/[pin]/[round]/+layout.server.js
@@ -1,15 +1,7 @@
-import { admin } from "$lib/admin.server"
-
 /** @type {import('./$types').LayoutServerLoad} */
 export async function load({ params }) {
-  const auctionSize = await admin
-    .database()
-    .ref(`auctions/${params.pin}/size`)
-    .get()
   return {
-    size: auctionSize.val(),
-    scores: [190, 200, 110, 90, 100, 200],
-    pin: params.pin,
+    scores: [890, 200, 110, 90, 100, 200],
     round: params.round,
   }
 }
